refactor: 都道府県リストの取得の部分のコンポーネントの分離
可読性を高めるため
